refactor: moves authorization controller to v1
